pyproject.toml
[build-system]
requires = ["setuptools>=42.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "llamadocx"
version = "0.1.0"
description = "A powerful toolkit for working with Microsoft Word documents"
readme = "README.md"
requires-python = ">=3.8"
license = {file = "LICENSE"}
authors = [
    {name = "LlamaDocx Team", email = "info@llamasearch.ai"}
]
maintainers = [
    {name = "LlamaDocx Team", email = "info@llamasearch.ai"}
]
keywords = ["docx", "word", "document", "template", "conversion"]
classifiers = [
    "Development Status :: 4 - Beta",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.8",
    "Programming Language :: Python :: 3.9",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Topic :: Office/Business",
    "Topic :: Software Development :: Libraries",
]
dependencies = [
    "python-docx>=0.8.11",
    "markdown>=3.4.1",
    "beautifulsoup4>=4.11.1",
]

[project.urls]
"Homepage" = "https://github.com/llamasearch/llamadocx"
"Bug Tracker" = "https://github.com/llamasearch/llamadocx/issues"
"Documentation" = "https://llamadocx.readthedocs.io/"

[project.optional-dependencies]
dev = [
    "pytest>=7.0.0",
    "pytest-cov>=3.0.0",
    "black>=22.1.0",
    "isort>=5.10.1",
    "flake8>=4.0.1",
    "mypy>=0.931",
]
convert = [
    "pypandoc>=1.6.0",
    "docx2pdf>=0.1.8",
]
full = [
    "pypandoc>=1.6.0",
    "docx2pdf>=0.1.8",
    "rapidfuzz>=2.0.0",
    "Pillow>=9.0.0",
    "opencv-python>=4.5.0",
]

[project.scripts]
llamadocx = "llamadocx.cli:cli"

[tool.setuptools]
package-dir = {"" = "src"}

[tool.setuptools.packages.find]
where = ["src"]

[tool.black]
line-length = 88
target-version = ["py38"]
include = '\.pyi?$'

[tool.isort]
profile = "black"
line_length = 88
multi_line_output = 3

[tool.mypy]
python_version = "3.8"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true
disallow_incomplete_defs = true

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = "test_*.py"
python_functions = "test_*" # Version bump for first release
